Updated templates

+++ terminusgps_payments/views/address_profiles.py
@@ -22,7 +22,7 @@ class AddressProfileCreateView(
     LoginRequiredMixin, HtmxTemplateResponseMixin, FormView
 ):
     content_type = "text/html"
-    extra_context = {"title": "Create Payment Method"}
+    extra_context = {"title": "Create Address Profile"}
     form_class = AddressProfileCreationForm
     http_method_names = ["get", "post"]
     partial_template_name = (
@@ -37,14 +37,19 @@ class AddressProfileCreateView(
         customer_profile = CustomerProfile.objects.get(user=self.request.user)
         address_profile = AddressProfile(customer_profile=customer_profile)
 
-        address = form.cleaned_data["address"]
-        default = form.cleaned_data["default"]
-
         try:
             service = AddressProfileService()
-            address_profile = service.create(
-                address_profile, address=address, default=default
+            address_profile, api_response = service.create(
+                address_profile,
+                address=form.cleaned_data["address"],
+                default=form.cleaned_data["default"],
             )
+            if api_response is None:
+                raise AuthorizenetControllerExecutionError(
+                    code="1",
+                    message="Whoops! Something went wrong with the Authorizenet API. Please try again later.",
+                )
+            address_profile.pk = int(api_response.customerAddressId)
             address_profile.save()
             return HttpResponse(
                 status=200,
@@ -81,17 +86,22 @@ class AddressProfileDetailView(
     raise_exception = False
     template_name = "terminusgps_payments/address_profiles/detail.html"
 
+    def get_context_data(self, **kwargs) -> dict[str, typing.Any]:
+        context: dict[str, typing.Any] = super().get_context_data(**kwargs)
+        if address_profile := kwargs.get("object"):
+            if address_profile.address is None:
+                # Retrieve address data from Authorizenet
+                service = AddressProfileService()
+                address_profile, api_response = service.get(address_profile)
+                if api_response is not None:
+                    address_profile.address = getattr(api_response, "address")
+                    address_profile.save()
+        return context
+
     def get_queryset(self) -> QuerySet:
         return AddressProfile.objects.for_user(
             self.request.user
         ).select_related("customer_profile")
-
-    def get_context_data(self, **kwargs) -> dict[str, typing.Any]:
-        context: dict[str, typing.Any] = super().get_context_data(**kwargs)
-        if address_profile := kwargs.get("object"):
-            service = AddressProfileService()
-            context["profile"] = service.get(address_profile)
-        return context
 
 
 class AddressProfileDeleteView(
@@ -114,19 +124,16 @@ class AddressProfileDeleteView(
             self.request.user
         ).select_related("customer_profile")
 
-    def get_context_data(self, **kwargs) -> dict[str, typing.Any]:
-        context: dict[str, typing.Any] = super().get_context_data(**kwargs)
-        if address_profile := kwargs.get("object"):
-            service = AddressProfileService()
-            context["profile"] = service.get(address_profile)
-        return context
-
     def form_valid(self, form: forms.ModelForm) -> HttpResponse:
         try:
             service = AddressProfileService()
             address_profile = self.get_object()
-
-            address_profile = service.delete(address_profile)
+            address_profile, api_response = service.delete(address_profile)
+            if api_response is None:
+                raise AuthorizenetControllerExecutionError(
+                    code="1",
+                    message="Whoops! Something went wrong calling the Authorizenet API. Please try again later.",
+                )
             address_profile.delete()
             return HttpResponse(
                 status=200,
